+++ api/auth/signer/v4/stream.go
@@ -5,25 +5,21 @@ import (
 	"strings"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws/credentials"
+	"github.com/aws/aws-sdk-go-v2/aws"
 )
-
-type credentialValueProvider interface {
-	Get() (credentials.Value, error)
-}
 
 // StreamSigner implements signing of event stream encoded payloads.
 type StreamSigner struct {
 	region  string
 	service string
 
-	credentials credentialValueProvider
+	credentials aws.Credentials
 
 	prevSig []byte
 }
 
 // NewStreamSigner creates a SigV4 signer used to sign Event Stream encoded messages.
-func NewStreamSigner(region, service string, seedSignature []byte, credentials *credentials.Credentials) *StreamSigner {
+func NewStreamSigner(region, service string, seedSignature []byte, credentials aws.Credentials) *StreamSigner {
 	return &StreamSigner{
 		region:      region,
 		service:     service,
@@ -34,12 +30,7 @@ func NewStreamSigner(region, service string, seedSignature []byte, credentials *
 
 // GetSignature takes an event stream encoded headers and payload and returns a signature.
 func (s *StreamSigner) GetSignature(headers, payload []byte, date time.Time) ([]byte, error) {
-	credValue, err := s.credentials.Get()
-	if err != nil {
-		return nil, err
-	}
-
-	sigKey := deriveSigningKey(s.region, s.service, credValue.SecretAccessKey, date)
+	sigKey := deriveSigningKey(s.region, s.service, s.credentials.SecretAccessKey, date)
 
 	keyPath := buildSigningScope(s.region, s.service, date)